# openapi.yaml
openapi: 3.1.0
info:
  title: NeroCD API
  version: 0.1.0
  description: |
    Public REST API for the NeroCD automation control plane.

    All public routes are versioned under `/api/v1`. List responses use a stable
    `{ "items": [...] }` envelope. Error responses use
    `{ "code": "stable_code", "error": "message" }`.

    List endpoints return a stable `{ "items": [...] }` envelope plus
    pagination metadata: `limit`, `offset`, `count`, and `total`. Clients can
    request `limit` and `offset` query parameters. Filtering is documented per
    endpoint. Mutations consume JSON request bodies and return the created
    resource or validation errors.
servers:
  - url: http://127.0.0.1:8080
    description: Local development server
security:
  - bearerAuth: []
  - browserSession: []
paths:
...
  /api/v1/templates:
    get:
      summary: List task templates
      operationId: listTemplates
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: Task template list. When `project_id` is supplied, only templates in that project are returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskTemplateListResponse"
        "401":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
    post:
      summary: Create task template
      operationId: createTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateInput"
      responses:
        "201":
          description: Created task template.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskTemplate"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
    patch:
      summary: Update task template
      operationId: updateTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - type: object
                  required: [id]
                  properties:
                    id:
                      type: string
                - $ref: "#/components/schemas/TemplateInput"
      responses:
        "200":
          description: Updated task template.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskTemplate"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: NeroCD session token
      description: CLI session and API token credential.
    browserSession:
      type: apiKey
      in: cookie
      name: nerocd_session
      description: "Browser-only local session cookie. Unsafe cookie-authenticated requests require X-NeroCD-CSRF: 1."
  parameters:
    ProjectID:
      name: project_id
      in: query
      required: false
      description: Optional project identifier filter.
      schema:
        type: string
...
  responses:
    Error:
      description: Stable error envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, error]
      properties:
        code:
          type: string
          enum: [unauthenticated, forbidden, not_found, bad_request, internal_error]
        error:
          type: string
...
    TaskTemplate:
      type: object
      required: [id, project_id, name, kind, run_spec, workflow, runner_tags, requires_ack]
      properties:
        id:
          type: string
        project_id:
          type: string
        name:
          type: string
        kind:
          type: string
        run_spec:
          $ref: "#/components/schemas/RunSpec"
        workflow:
          $ref: "#/components/schemas/Workflow"
        runner_tags:
          type: array
          items:
            type: string
        requires_ack:
          type: boolean
    TemplateInput:
      type: object
      required: [project_id, name, run_spec]
      properties:
        project_id:
          type: string
        name:
          type: string
        kind:
          type: string
        run_spec:
          $ref: "#/components/schemas/RunSpec"
        workflow:
          $ref: "#/components/schemas/Workflow"
        runner_tags:
          type: array
          items:
            type: string
        requires_ack:
          type: boolean
    TaskTemplateListResponse:
      type: object
      required: [items, limit, offset, count, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TaskTemplate"
        limit:
          type: integer
          minimum: 0
        offset:
          type: integer
          minimum: 0
        count:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
...
    RunSpec:
      type: object
      required: [type, inputs]
      properties:
        type:
          type: string
        inputs:
          type: object
          additionalProperties: true
        repository:
          $ref: "#/components/schemas/RepositoryRef"
        process:
          $ref: "#/components/schemas/ProcessSpec"
        artifacts:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactSpec"
        secrets:
          type: array
          items:
            $ref: "#/components/schemas/SecretBinding"
        workflow:
          $ref: "#/components/schemas/Workflow"
    RepositoryRef:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
        provider:
          type: string
        ref:
          type: string
        path:
          type: string
    ProcessSpec:
      type: object
      required: [command]
      properties:
        command:
          type: array
          items:
            type: string
        working_dir:
          type: string
        environment:
          type: object
          additionalProperties:
            type: string
        timeout_seconds:
          type: integer
    ArtifactSpec:
      type: object
      required: [name, path, required]
      properties:
        name:
          type: string
        path:
          type: string
        required:
          type: boolean
        retention:
          type: string
    SecretBinding:
      type: object
      required: [name, provider, reference, target, required]
      properties:
        name:
          type: string
        provider:
          type: string
        reference:
          type: string
        target:
          type: string
        required:
          type: boolean
        version:
          type: string
        fingerprint:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        redact_encodings:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [base64, base64url, hex]
        classification:
          type: string
          enum: [development]
...
    Workflow:
      type: object
      required: [steps]
      properties:
        steps:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowStep"
    WorkflowStep:
      type: object
      required: [id, name, run_spec, requires_ack]
      properties:
        id:
          type: string
        name:
          type: string
        run_spec:
          $ref: "#/components/schemas/RunSpec"
        depends_on:
          type: array
          items:
            type: string
        requires_ack:
          type: boolean
...